requirements.txt
APScheduler==3.10.4
asqlite==2.0.0
asyncpg==0.29.0
beautifulsoup4==4.12.3
discord.py==2.4.0
lxml==5.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
googletrans==3.1.0a0
